test: fix EntityScoreCardTable-TestWithData title test
- Title test in EntityScoreCardTable-TestWithData did not actually check
  the title was present.
- Prior version of test triggered `act()` warning

```
console.error
Warning: An update to ScoreCardTable inside a test was not wrapped in act(...).
When testing, code that causes React state updates should be wrapped into act(...):

act(() => {
  /* fire events that update state */
});
/* assert on the output */

This ensures that you're testing the behavior the user would see in the browser. Learn more at https://reactjs.org/link/wrap-tests-with-act
```

Signed-off-by: Robin H. Johnson <rjohnson@coreweave.com>